Surya Murali's DSAPP RA Project
In this exercise, I have analyzed data from KDD Cup 2014 porject in Kaggle to finding exciting K12 student projects for donors. In addition to developing a predictive model, I have explored the data initially and have come up with a few ideas or ways to improve the project submission system, the evaluation criteria and the school administrations as it currently exists. As a data scientist, I consider working on social problems to be most noble, hence I enjoyed working on this problem.